refactor(notice): move attachments to system downloads, add management page
Move notice parts into jwc/ subdirectory. Switch attachment download path
from app documents to system Downloads/Bugaoshan/. Add downloaded attachments
management page with open/share/delete/search/sort.
