+++ lib/pages/campus/notice/jwc/notice_models.dart
