+++ lib/utils/dock_utils.dart
@@ -4,7 +4,7 @@ import 'package:bugaoshan/pages/campus/academic_calendar/academic_calendar_page.
 import 'package:bugaoshan/pages/campus/balance_query/balance_query_page.dart';
 import 'package:bugaoshan/pages/campus/classroom/classroom_page.dart';
 import 'package:bugaoshan/pages/campus/fitness_test/fitness_test_page.dart';
-import 'package:bugaoshan/pages/campus/notice/campus_notice_page.dart';
+import 'package:bugaoshan/pages/campus/notice/jwc/campus_notice_page.dart';
 import 'package:bugaoshan/pages/campus/ccyl/ccyl_page.dart';
 import 'package:bugaoshan/pages/campus/grades/grades_page.dart';
 import 'package:bugaoshan/pages/campus/network_device/network_device_page.dart';
